requirements.txt
Flask
python-dotenv
python-docx
PyMuPDF
sentence-transformers
faiss-cpu
numpy
openai
pytest
